ig_vendor_py/pyproject.toml
[project]
name = "ig-vendor-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "instagrapi>=2.1.3",
    "peewee>=3.17.9",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
